[Feature] Implement user settings to be saved as a diff object.
To edit options, we allow users to edit the given JSON. Earlier, we were saving the entire user settings JSON (in localStorage), and overriding it with the default settings to calculate the final settings. This was leading to problematic scenarios where users could continue to have obselete settings (when default settings were changed over different versions of the extension).
Now, we save only the user settings diff (in localStorage). That is, we only save what the user has changed. We also keep track of changes in arrays, by keeping record of elements that are added or removed.
To keep things simple and predictable, urlDataMap settings (URL/site specific settings) cannot be changed for now. Only general settings can be modified.