Revert "[Feature] Commented out lines no longer required (based on commit a4648a0)"
This reverts commit c99151cc0fe1d3ee08e5eab8667481e740e0699b.
The lines that had been commented out are still useful -- they
give the browser a chance to make visible some element if it's occluded
etc (even if it was already in the viewport). Not sure if the browsers
do something like this, but they might or might in the future.