[Minor] changed comments
(was trying to use  "pointer-events: stroke" to allow the hovered overlay to trigger a mouseover event which we  could directly use, to call dehover(). But "stroke" seemed to be not allowing clicks etc from within the overlay to be going through as well, hence couldn't be used. TODO: try this later)