[Fix] In mod_CUsMgr, reset urlData related global references in reset()
Fixes #50. Issue: The overlay was not going away on URL change. The issue is specific to the scenario: In a single page app, when the URL changes before the DOM, and the new URL does not have a CUsSpecifier associated and the previous one does. In this case, the module variable CUsSpecifier was not reset correctly, and due to that, CUs were incorrectly being found during DOM mutations.

What was going on:
- In this case on Github, the URL changes before the DOM has (entirely) changed.
- On URL change, we call the reset() and setup() of mod_CUsMgr. These were working as expected. Also, setup() returns early since there is no CUsSpecifier for the current URL.
- Soon after, some dom_mutation events occur. This is where the issue is. Since the CUsSpecifier reference was not reset (in either reset() or setup()), and since DOM has not updated yet, there still are CUs on the page. And the overlay gets rendered for the first CU.
- When the DOM has completely changed for the new URL, the overlay still remains.